[TeamC] move question generate form into views
